# openapi/openapi.yaml
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact: {}
  description: Backend control-plane API for the adult English voice-room product.
  title: Slogan Voice Room API
  version: 0.0.1
openapi: 3.0.0
paths:
  /v1/auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      responses:
        "204":
          description: ""
        "401":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
      security:
        - bearer: []
      tags:
        - auth
  /v1/auth/oauth/{provider}/exchange:
    post:
      operationId: AuthController_exchange
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            enum:
              - google
              - wechat
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                authorizationCode:
                  maxLength: 4096
                  minLength: 1
                  type: string
                codeVerifier:
                  maxLength: 128
                  minLength: 43
                  type: string
                deviceName:
                  maxLength: 120
                  type: string
                redirectUri:
                  example: slogan://oauth/google
                  type: string
              required:
                - authorizationCode
                - redirectUri
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  created:
                    type: boolean
                  onboardingState:
                    enum:
                      - PROFILE_REQUIRED
                      - AGE_RESTRICTED
                      - ELIGIBLE
                    type: string
                  suggestedProfile:
                    properties:
                      avatarUrl:
                        type: string
                      displayName:
                        type: string
                    type: object
                  tokens:
                    properties:
                      accessToken:
                        type: string
                      accessTokenExpiresInSeconds:
                        example: 900
                        type: number
                      refreshToken:
                        type: string
                      refreshTokenExpiresAt:
                        format: date-time
                        type: string
                    required:
                      - accessToken
                      - accessTokenExpiresInSeconds
                      - refreshToken
                      - refreshTokenExpiresAt
                    type: object
                  userId:
                    format: uuid
                    type: string
                required:
                  - userId
                  - created
                  - onboardingState
                  - tokens
                type: object
          description: ""
        "400":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "429":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "503":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "504":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
      tags:
        - auth
  /v1/auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                refreshToken:
                  maxLength: 200
                  minLength: 40
                  type: string
              required:
                - refreshToken
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  accessToken:
                    type: string
                  accessTokenExpiresInSeconds:
                    example: 900
                    type: number
                  refreshToken:
                    type: string
                  refreshTokenExpiresAt:
                    format: date-time
                    type: string
                required:
                  - accessToken
                  - accessTokenExpiresInSeconds
                  - refreshToken
                  - refreshTokenExpiresAt
                type: object
          description: ""
        "400":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "401":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "429":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
      tags:
        - auth
...
servers: []
...